added `recovery-code` filtering for get Authenticator
